job_agent/requirements.txt
playwright
requests
beautifulsoup4
apscheduler
flask
anthropic
python-dotenv
plyer
pyyaml
